Fix Card box showing Welcome information once the user has logged in, this should have just been for the login or register section so moved this to be in the actual partial instead of around the partial call in the home.
Also moved Welcome message to the top of the sidebar, I need to make it so it takes the users name. Either have the first log in take the user to an account set up page or leave it as an email, this might cause issues with long emails that I cba to look into.